Fix broken cluster tests caused by uninitialized metrics
